Removed a reminder to describe clipboard artifacts repro case.
I'll put them here:

1. Open MrClippy.
1. Save your clipboard as a clip file.
1. Copy change your clipboard (just copy anything else).
1. Open file saved in previous steps.
1. Write it into the clipboard using "Write" button.

Now if you check for intance HTML type, you'll see it contains a lot of garbage. In my case it was full content of the JSON stored by the .clip file. It sounds like a reasonable assumption that the data of HTML Format in loaded snapshot is just a view to entire loaded JSON blob, with a specific offsets, just to pick the value - but I don't want to really to dive that deep right now.