Extracted snapshot storage key creation to a separate function.
That fixes a case where persistent storage key could be used by more than one entry, leading to unexpected results. Here's my todo note describing this case:

"There is still a bug that the key (perist storage) is not modified when removing. So if you remove an element that had index 0, there are items with index 1, 2, 3 remainaing.

Since there's a simplified logic for index obtaining (it just uses items.length) it becomes 3 again and it overwrites the last item."